[127] ContactScreen refactored to be 'HomeScreen'
